Remove int type-hint from $rootId in GetSearchablePagesListener.php
The $rootId parameter was type-hinted to int|null. But the SitemapController from contao does pass a string variable so an exception was thrown.
Because of compatibility with 7.4, we can't use union type hints, so it has to be removed. Updated the PHP documentation block to typehint for int|string|null.
If a string is passed as $rootId parameter, it will be converted to an int if possible. If this cast fails, it will be set to null;